minimodem: --tx SAME emits 16 sync-byte preamble
Emit a preamble of 16 (hardcoded) sync-bytes if --sync-byte is
specified with --tx mode.  This makes --tx SAME emit the required
preamble of 16 0xAB's automatically.  So --rx SAME now directly
decodes the output of --tx SAME without needing to manually inject
the preamble byte(s), simplifying the test case.

Thanks to Harold Giddings, for reporting the issue.